Reorganize and remove reports folder
