+++ kolibri/plugins/coach/assets/src/constants/index.js
@@ -15,6 +15,10 @@ export const PageNames = {
   QUIZ_SELECT_RESOURCES: 'QUIZ_SELECT_RESOURCES',
   QUIZ_SELECT_PRACTICE_QUIZ: 'QUIZ_SELECT_PRACTICE_QUIZ',
   QUIZ_SECTION_ORDER: 'QUIZ_SECTION_ORDER',
+  QUIZ_LEARNER_REPORT: 'QUIZ_LEARNER_REPORT',
+  QUIZ_PREVIEW: 'QUIZ_PREVIEW',
+  QUIZ_QUESTION_PAGE_ROOT: 'QUIZ_QUESTION_PAGE_ROOT',
+  QUIZ_QUESTION_REPORT: 'QUIZ_QUESTION_REPORT',
   BOOK_MARKED_RESOURCES: 'BOOK_MARKED_RESOURCES',
 
   EXAM_REPORT: 'EXAM_REPORT',
@@ -24,20 +28,14 @@ export const PageNames = {
   LESSON_EXERCISE_LEARNER_REPORT: 'LESSON_EXERCISE_LEARNER_REPORT',
   REPORTS_GROUP_REPORT_LESSON_EXERCISE_LEARNER_PAGE_ROOT:
     'REPORTS_GROUP_REPORT_LESSON_EXERCISE_LEARNER_PAGE_ROOT',
-  REPORTS_LEARNER_ACTIVITY_EXERCISE_PAGE_ROOT: 'REPORTS_LEARNER_ACTIVITY_EXERCISE_PAGE_ROOT',
-  REPORTS_LEARNER_REPORT_LESSON_EXERCISE_PAGE_ROOT:
-    'REPORTS_LEARNER_REPORT_LESSON_EXERCISE_PAGE_ROOT',
   LESSON_LEARNER_EXERCISE_PAGE_ROOT: 'LESSON_LEARNER_EXERCISE_PAGE_ROOT',
   LESSON_LEARNER_EXERCISE_REPORT: 'LESSON_LEARNER_EXERCISE_REPORT',
   REPORTS_GROUP_LEARNER_REPORT_QUIZ_PAGE_ROOT: 'REPORTS_GROUP_LEARNER_REPORT_QUIZ_PAGE_ROOT',
-  REPORTS_GROUP_REPORT_QUIZ_LEARNER_PAGE_ROOT: 'REPORTS_GROUP_REPORT_QUIZ_LEARNER_PAGE_ROOT',
+  QUIZ_LEARNER_PAGE_ROOT: 'QUIZ_LEARNER_PAGE_ROOT',
   REPORTS_LEARNER_REPORT_QUIZ_PAGE_ROOT: 'REPORTS_LEARNER_REPORT_QUIZ_PAGE_ROOT',
   REPORTS_QUIZ_LEARNER_PAGE_ROOT: 'REPORTS_QUIZ_LEARNER_PAGE_ROOT',
   LESSON_EXERCISE_QUESTION_PAGE_ROOT: 'LESSON_EXERCISE_QUESTION_PAGE_ROOT',
-  REPORTS_GROUP_REPORT_LESSON_EXERCISE_QUESTION_PAGE_ROOT:
-    'REPORTS_GROUP_REPORT_LESSON_EXERCISE_QUESTION_PAGE_ROOT',
   REPORTS_GROUP_REPORT_QUIZ_QUESTION_PAGE_ROOT: 'REPORTS_GROUP_REPORT_QUIZ_QUESTION_PAGE_ROOT',
-  REPORTS_QUIZ_QUESTION_PAGE_ROOT: 'REPORTS_QUIZ_QUESTION_PAGE_ROOT',
 
   /* Lessons */
   LESSON_CREATION_ROOT: 'LESSON_CREATION_ROOT',
@@ -68,6 +66,7 @@ export const PageNames = {
 
   LEARNERS_ROOT: 'LEARNERS_ROOT',
   LEARNER_SUMMARY: 'LEARNER_SUMMARY',
+  LEARNER_LESSON_REPORT: 'LEARNER_LESSON_REPORT',
 
   GROUPS_ROOT: 'GROUPS_ROOT',
   GROUP_SUMMARY: 'GROUP_SUMMARY',
@@ -76,6 +75,7 @@ export const PageNames = {
   GROUP_LESSON_LEARNER: 'GROUP_LESSON_LEARNER',
   GROUP_LESSON_EXERCISE_QUESTIONS_REPORT: 'GROUP_LESSON_EXERCISE_QUESTIONS_REPORT',
   GROUP_LESSON_EXERCISE_LEARNER_REPORT: 'GROUP_LESSON_EXERCISE_LEARNER_REPORT',
+  GROUP_EXAM_SUMMARY: 'GROUP_EXAM_SUMMARY',
 };
 
 export const GroupModals = {
