+++ kolibri/plugins/coach/assets/src/constants/lastPagesConstants.js
@@ -1,7 +1,6 @@
 export const LastPages = {
   HOME_PAGE: 'homepage',
   HOME_ACTIVITY: 'homeactivity',
-  GROUP_ACTIVITY: 'groupactivity',
   LEARNER_ACTIVITY: 'learneractivity',
   EXERCISE_LEARNER_LIST: 'exerciselearnerlist',
   EXERCISE_LEARNER_LIST_BY_GROUPS: 'exerciselearnerlistbygroups',
