--- kolibri/plugins/coach/assets/src/views/lessons/reports/LessonExerciseQuestionPage.vue
@@ -1,55 +1,0 @@
-<template>
-
-  <CoachImmersivePage
-    :appBarTitle="title"
-    icon="back"
-    :pageTitle="title"
-    :primary="false"
-    :route="toolbarRoute"
-  >
-    <QuestionLearnersReport @navigate="handleNavigation" />
-  </CoachImmersivePage>
-
-</template>
-
-
-<script>
-
-  import { mapState } from 'vuex';
-  import commonCoach from '../../common';
-  import CoachImmersivePage from '../../CoachImmersivePage';
-  import QuestionLearnersReport from '../../common/QuestionLearnersReport';
-  import { PageNames } from '../../../constants';
-
-  export default {
-    name: 'LessonExerciseQuestionPage',
-    components: {
-      CoachImmersivePage,
-      QuestionLearnersReport,
-    },
-    mixins: [commonCoach],
-    computed: {
-      ...mapState('questionDetail', ['title']),
-      toolbarRoute() {
-        const backRoute = this.backRouteForQuery(this.$route.query);
-        return backRoute || this.classRoute(PageNames.LESSON_EXERCISE_QUESTIONS_REPORT, {});
-      },
-    },
-    methods: {
-      handleNavigation(params) {
-        this.$router.push({
-          name: this.name,
-          params: {
-            classId: this.$route.params.classId,
-            lessonId: this.$route.params.lessonId,
-            ...params,
-          },
-        });
-      },
-    },
-  };
-
-</script>
-
-
-<style lang="scss" scoped></style>
