--- kolibri/plugins/coach/assets/src/views/reports/ReportsGroupReportLessonExerciseQuestionPage.vue
@@ -1,57 +1,0 @@
-<template>
-
-  <CoachImmersivePage
-    :appBarTitle="exercise.title"
-    icon="back"
-    :pageTitle="title"
-    :primary="false"
-    :route="toolbarRoute"
-  >
-    <QuestionLearnersReport @navigate="handleNavigation" />
-  </CoachImmersivePage>
-
-</template>
-
-
-<script>
-
-  import { mapState } from 'vuex';
-  import commonCoach from '../common';
-  import CoachImmersivePage from '../CoachImmersivePage';
-  import QuestionLearnersReport from '../common/QuestionLearnersReport';
-  import { PageNames } from '../../constants';
-
-  export default {
-    name: 'ReportsGroupReportLessonExerciseQuestionPage',
-    components: {
-      CoachImmersivePage,
-      QuestionLearnersReport,
-    },
-    mixins: [commonCoach],
-    computed: {
-      ...mapState('questionDetail', ['title']),
-      ...mapState('exerciseDetail', ['exercise']),
-      toolbarRoute() {
-        const backRoute = this.backRouteForQuery(this.$route.query);
-        return backRoute || this.classRoute(PageNames.GROUP_LESSON_EXERCISE_QUESTIONS_REPORT, {});
-      },
-    },
-    methods: {
-      handleNavigation(params) {
-        this.$router.push({
-          name: this.name,
-          params: {
-            classId: this.$route.params.classId,
-            groupId: this.$route.params.groupId,
-            lessonId: this.$route.params.lessonId,
-            ...params,
-          },
-        });
-      },
-    },
-  };
-
-</script>
-
-
-<style lang="scss" scoped></style>
