--- kolibri/plugins/coach/assets/src/views/reports/ReportsGroupReportQuizQuestionListPage.vue
@@ -1,19 +1,0 @@
-<template>
-
-  <QuizQuestionListPageBase />
-
-</template>
-
-
-<script>
-
-  import QuizQuestionListPageBase from './QuizQuestionListPageBase';
-
-  export default {
-    name: 'ReportsGroupReportQuizQuestionListPage',
-    components: {
-      QuizQuestionListPageBase,
-    },
-  };
-
-</script>
