--- kolibri/plugins/coach/assets/src/views/reports/ReportsLearnerActivityExercisePage.vue
@@ -1,51 +1,0 @@
-<template>
-
-  <CoachImmersivePage
-    :appBarTitle="exercise.title"
-    :pageTitle="exercise.title"
-    :route="toolbarRoute"
-  >
-    <LearnerExerciseReport @navigate="handleNavigation" />
-  </CoachImmersivePage>
-
-</template>
-
-
-<script>
-
-  import { mapState } from 'vuex';
-  import commonCoach from '../common';
-  import CoachImmersivePage from '../CoachImmersivePage';
-  import LearnerExerciseReport from '../common/LearnerExerciseReport';
-
-  export default {
-    name: 'ReportsLearnerActivityExercisePage',
-    components: {
-      CoachImmersivePage,
-      LearnerExerciseReport,
-    },
-    mixins: [commonCoach],
-    computed: {
-      ...mapState('exerciseDetail', ['exercise']),
-      toolbarRoute() {
-        const backRoute = this.backRouteForQuery(this.$route.query);
-        return backRoute || this.classRoute('ReportsLearnerActivityPage', {});
-      },
-    },
-    methods: {
-      handleNavigation(params) {
-        this.$router.push({
-          name: this.name,
-          params: {
-            classId: this.$route.params.classId,
-            ...params,
-          },
-        });
-      },
-    },
-  };
-
-</script>
-
-
-<style lang="scss" scoped></style>
