--- kolibri/plugins/coach/assets/src/views/reports/ReportsQuizQuestionListPage.vue
@@ -1,19 +1,0 @@
-<template>
-
-  <QuizQuestionListPageBase />
-
-</template>
-
-
-<script>
-
-  import QuizQuestionListPageBase from './QuizQuestionListPageBase';
-
-  export default {
-    name: 'ReportsQuizQuestionListPage',
-    components: {
-      QuizQuestionListPageBase,
-    },
-  };
-
-</script>
