--- kolibri/plugins/coach/assets/src/views/reports/allReportsPages.js
@@ -1,41 +1,0 @@
-import ReportsGroupActivityPage from './ReportsGroupActivityPage';
-import ReportsGroupLearnerListPage from './ReportsGroupLearnerListPage';
-import ReportsGroupReportLessonExerciseLearnerPage from './ReportsGroupReportLessonExerciseLearnerPage';
-import ReportsGroupReportLessonExerciseQuestionPage from './ReportsGroupReportLessonExerciseQuestionPage';
-import ReportsGroupReportLessonResourceLearnerListPage from './ReportsGroupReportLessonResourceLearnerListPage';
-import ReportsGroupReportPage from './ReportsGroupReportPage';
-import ReportsGroupReportQuizLearnerListPage from './ReportsGroupReportQuizLearnerListPage';
-import ReportsGroupReportQuizLearnerPage from './ReportsGroupReportQuizLearnerPage';
-import ReportsGroupReportQuizQuestionListPage from './ReportsGroupReportQuizQuestionListPage';
-import ReportsGroupReportQuizQuestionPage from './ReportsGroupReportQuizQuestionPage';
-import ReportsLearnerActivityExercisePage from './ReportsLearnerActivityExercisePage';
-import ReportsLearnerReportLessonExercisePage from './ReportsLearnerReportLessonExercisePage';
-import ReportsLearnerReportLessonPage from './ReportsLearnerReportLessonPage';
-import ReportsLearnerReportQuizPage from './ReportsLearnerReportQuizPage';
-import ReportsQuizLearnerListPage from './ReportsQuizLearnerListPage';
-import ReportsQuizLearnerPage from './ReportsQuizLearnerPage';
-import ReportsQuizPreviewPage from './ReportsQuizPreviewPage';
-import ReportsQuizQuestionListPage from './ReportsQuizQuestionListPage';
-import ReportsQuizQuestionPage from './ReportsQuizQuestionPage';
-
-export default {
-  ReportsGroupActivityPage,
-  ReportsGroupLearnerListPage,
-  ReportsGroupReportLessonExerciseLearnerPage,
-  ReportsGroupReportLessonExerciseQuestionPage,
-  ReportsGroupReportLessonResourceLearnerListPage,
-  ReportsGroupReportPage,
-  ReportsGroupReportQuizLearnerListPage,
-  ReportsGroupReportQuizLearnerPage,
-  ReportsGroupReportQuizQuestionListPage,
-  ReportsGroupReportQuizQuestionPage,
-  ReportsLearnerActivityExercisePage,
-  ReportsLearnerReportLessonExercisePage,
-  ReportsLearnerReportLessonPage,
-  ReportsLearnerReportQuizPage,
-  ReportsQuizLearnerListPage,
-  ReportsQuizLearnerPage,
-  ReportsQuizPreviewPage,
-  ReportsQuizQuestionListPage,
-  ReportsQuizQuestionPage,
-};
